Revert "Feature/insurance card image preprocessing (#8729)"
This reverts commit a681e1c3de89e7c5bfe9e4a693f840a46b876200.
