+++ apps/ehr/src/api/api.ts
@@ -205,8 +205,6 @@ import {
   RenameCustomFolderOutput,
   RetryActionLogInput,
   RetryActionLogOutput,
-  RotateInsuranceCardImageInput,
-  RotateInsuranceCardImageResponse,
   SaveFollowupEncounterZambdaInput,
   SaveFollowupEncounterZambdaOutput,
   SaveRadiologyReportZambdaInput,
@@ -334,7 +332,6 @@ const VISIT_DETAILS_TO_PDF_ZAMBDA_ID = 'visit-details-to-pdf';
 const PENDING_SUPERVISOR_APPROVAL_ZAMBDA_ID = 'pending-supervisor-approval';
 const SEND_RECEIPT_BY_EMAIL_ZAMBDA_ID = 'send-receipt-by-email';
 const BULK_UPDATE_INSURANCE_STATUS_ZAMBDA_ID = 'bulk-update-insurance-status';
-const ROTATE_INSURANCE_CARD_IMAGE_ZAMBDA_ID = 'rotate-insurance-card-image';
 const ADMIN_GET_QUICK_PICKS_ZAMBDA_ID = 'admin-get-quick-picks';
 const ADMIN_CREATE_QUICK_PICK_ZAMBDA_ID = 'admin-create-quick-pick';
 const ADMIN_UPDATE_QUICK_PICK_ZAMBDA_ID = 'admin-update-quick-pick';
@@ -1947,30 +1944,6 @@ export const bulkUpdateInsuranceStatus = async (
     }
     const response = await oystehr.zambda.execute({
       id: BULK_UPDATE_INSURANCE_STATUS_ZAMBDA_ID,
-      ...parameters,
-    });
-    return chooseJson(response);
-  } catch (error: unknown) {
-    console.log(error);
-    throw error;
-  }
-};
-
-/**
- * Rotates a stored insurance-card image `rotationDegrees` CLOCKWISE, server-side and in place
- * (same attachment url), so the fix persists everywhere the image is consumed. Also clears the
- * extraction's "looks rotated" (`readable=false`) hint.
- */
-export const rotateInsuranceCardImage = async (
-  oystehr: Oystehr,
-  parameters: RotateInsuranceCardImageInput
-): Promise<RotateInsuranceCardImageResponse> => {
-  try {
-    if (ROTATE_INSURANCE_CARD_IMAGE_ZAMBDA_ID == null) {
-      throw new Error('rotate insurance card image zambda environment variable could not be loaded');
-    }
-    const response = await oystehr.zambda.execute({
-      id: ROTATE_INSURANCE_CARD_IMAGE_ZAMBDA_ID,
       ...parameters,
     });
     return chooseJson(response);
