+++ packages/utils/lib/utils/index.ts
@@ -1,6 +1,5 @@
 export * from './date';
 export * from './file';
-export * from './image-orientation';
 export * from './objects';
 export * from './pdf';
 export * from './timeout';
