+++ packages/zambdas/package.json
@@ -85,7 +85,6 @@
     "diacritics": "^1.3.0",
     "fast-json-patch": "^3.1.1",
     "handlebars": "^4.7.8",
-    "jimp": "^1.6.1",
     "luxon": "^3.4.3",
     "pdfmake": "^0.3.6",
     "query-string": "^8.1.0",
